requirement.txt
asgiref==3.5.0
black==22.3.0
camelcase==0.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.4
Django==4.0.3
django-filter==21.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-safedelete==1.1.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
filelock==3.6.0
greenlet==1.1.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy==0.910
mypy-extensions==0.4.3
mysql-connector==2.2.9
mysql-connector-python==8.0.28
mysqlclient==2.1.0
openapi-codec==1.3.2
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
protobuf==4.21.1
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytz==2022.1
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
SQLAlchemy==1.4.32
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.1.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.13.4
virtualenvwrapper-win==1.2.7
